use bool variables instead of connecting/disconnecting signals
also move simple QgsVertexEntry to qgsvertexeditor.h
